chore: remove unused files and improve code structure

+++ codegen/llvm/src/codegen_llvm/expression/array.cpp
@@ -7,6 +7,7 @@
 llvm::Value* ArrayExpressionCodegen::codegen() const {
   auto llvm_array_type =
       LLVMCodegenUtils::type_to_llvm_type(context, expression->get_type());
+
   auto array_ptr = context->builder->CreateAlloca(llvm_array_type, nullptr);
 
   size_t idx = 0;
@@ -32,11 +33,11 @@ llvm::Value* ArrayExpressionCodegen::codegen() const {
 
     auto index_ptr = context->builder->CreateGEP(
         llvm_array_type, array_ptr,
-        {context->builder->getInt32(0), context->builder->getInt32(idx++)});
+        {context->builder->getInt32(0),
+         context->builder->getInt32(static_cast<unsigned>(idx++))});
 
     context->builder->CreateStore(llvm_value, index_ptr);
   }
 
-  // Return the loaded array value, not the pointer
   return context->builder->CreateLoad(llvm_array_type, array_ptr);
 }
